camera.py输出图像

+++ Inc/pid.h
@@ -10,9 +10,9 @@ extern void pid_init(void);
  * adjust when go straight with pid
  * @param int16_t error (positive for near left side now)
 *********/
-extern void straight_adjustment(int16_t error);
+extern void straight_adjustment(int32_t error);
 
 
-extern void straight_param_change(int16_t Kp, int16_t Ki, int16_t Kd);
+extern void straight_param_change(float Kp, float Ki, float Kd);
 
 #endif
